split up and breakup files and interfaces into reasonable files
