--- socket/src/commonMain/kotlin/mqtt/socket/ReadStats.kt
@@ -1,3 +1,0 @@
-package mqtt.socket
-
-expect fun readStats(port: UShort, contains: String): List<String>
