+++ socket/src/commonMain/kotlin/mqtt/socket/ServerSocket.kt
@@ -1,0 +1,25 @@
+package mqtt.socket
+
+import kotlin.time.ExperimentalTime
+
+
+@ExperimentalTime
+@ExperimentalUnsignedTypes
+interface ServerSocket : SuspendCloseable {
+    suspend fun bind(
+        port: UShort? = null,
+        host: String? = null,
+        socketOptions: SocketOptions? = null,
+        backlog: UInt = 0.toUInt()
+    ): SocketOptions
+
+    suspend fun accept(): ClientSocket
+    fun isOpen(): Boolean
+    fun port(): UShort?
+}
+
+@ExperimentalUnsignedTypes
+@ExperimentalTime
+expect fun asyncServerSocket(): ServerSocket
+
+expect fun readStats(port: UShort, contains: String): List<String>
