+++ socket/src/commonMain/kotlin/mqtt/socket/SuspendCloseable.kt
@@ -1,0 +1,5 @@
+package mqtt.socket
+
+interface SuspendCloseable {
+    suspend fun close()
+}
