+++ socket/src/jvmMain/kotlin/mqtt/socket/Buffer.kt
@@ -1,6 +1,5 @@
 package mqtt.socket
 
-import mqtt.socket.nio.JvmBuffer
 import java.io.RandomAccessFile
 import java.nio.ByteBuffer
 import java.nio.channels.FileChannel
