+++ socket/src/jvmMain/kotlin/mqtt/socket/nio2/AsyncBaseClientSocket.kt
@@ -1,8 +1,8 @@
 package mqtt.socket.nio2
 
+import mqtt.socket.JvmBuffer
 import mqtt.socket.PlatformBuffer
 import mqtt.socket.nio.ByteBufferClientSocket
-import mqtt.socket.nio.JvmBuffer
 import mqtt.socket.nio2.util.aRead
 import mqtt.socket.nio2.util.aWrite
 import mqtt.socket.nio2.util.assignedPort
